Updated edit_startup_arguments and edit_startup_item
Refactored the two functions so the option to save
the updated data is in the menu shown for
edit_startup_item. Also, updated the code in
edit_startup_item to update the argument count
when the user edits existing arguments and adds a
new one as well as when there are no arguments and
the user adds some. Finally, updated the
docstring for edit_startup_item to show that this
function returns a list containing the updated
startup items