Made changes to the user menus and how invalid choices are handled
Made a minor change to the use of newlines for
styling the message "Please enter a valid choice"
in the function user_menu_chooser. Also changed
the message to "That is an invalid choice!".

Tested this out with the main menu, the edit
startup items menu, the edit specific startup item
menu, the edit arguments menu, and the delete
arguments menu.