Made changes to edit_startup_arguments and started creating option to delete an argument
Removed the boolean variable create_menu so now
the menu is generated on every iteration of the
while loop. Moved up the code for argument
deletion, argument addition, and argument editing
to be elif blocks in the main if statement in the
while loop. Started coding in the argument
deletion logic. Currently, a menu displays with
the arguments listed and an option to cancel.
However, no argument is yet deleted.

Forgot to add to previous commit but changed the
else blocks in json_editor and
edit_startup_arguments to elif blocks based on if
the user_choice variable is greater than 0. This
is because the function user_menu_chooser will
return a 0 if the user enters an invalid choice.
By checking for this instead of using an else
block, the user can't get past a multiple-choice
menu unless they enter a valid option.