Finished coding ability to delete a startup argument
Along the way, moved the code to create the edit
startup item menu into the while loop. Changed the
code to use the variable new_args_list instead of
args_list because it's new_args_list that is
updated when adding or deleting an argument. Also,
updated the code to recalculate the variable
args_count after a deletion or addition by getting
the length of new_args_list.