Allowed ability to quit the program from any menu
Modified the function user_menu_chooser to add a
quit option to the menu passed in. Sys.exit is
used if the user chooses that option.

Also modified the main loop so there are only 3
options since quit is now handled elsewhere.

Tested the functionality from a few different
menus and it works.