Created new function to explain the program and a welcome message
Added a welcome message to the main menu with a
little tag line. Also, added a new option to the
main menu to explain the program and created a
function to do that. Updated the main menu
match-case statement. The function is called
explain_demord and currently just prints out:

Currently under construction...come back later...