modified the function signature for edit_startup_item
While planning out how to save the changed startup
item data, realized the helper function to save
the data will have to call the function
json_reader which takes in a dictionary and a
string to determine the full path of the
startup_data.json file.

After playing around with global variables, I
decided it would be easier to pass in the
variables as parameters since the function
json_editor, which calls edit_startup_item, takes
in those two variables as parameters.

So now, edit_startup_item takes in three
parameters: the startup item data as a dictionary,
the path to the JSON file as a list, and the name
of the JSON file as a string. Also adjusted the
call to edit_startup_item from json_editor to pass
in all 3 parameters.