Made two more changes to the function edit_startup_item
Added an input statement after the startup item
data is printed out, so it doesn't scroll off
screen. The statement says to press any key to
continue, which lets the user read the startup
item data in full before being presented with the
next menu.

Also, filled in case 6 in the edit_startup_item
menu for the match-block to call a new helper
function called save_startup_item. Decided the
helper function signature will be similar to the
functions json_reader or json_writer. This new
function will take in 3 parameters - a dictionary
of the startup item JSON data, a list representing
the path to the startup_data.json file, and a
string with the name of the startup_data.json
file. In addition, this function will return two
variables - a boolean for the status of the save
attempt and a string message of any error.