Started creating the helper function save_startup_item
Started the helper function to save a startup
item. As mentioned in commit 6a5bb8b, the
function takes in 3 parameters and returns 2
variables.

Added the docstring for the function, but
currently, it only prints out the startup_item
dictionary and returns a hardcoded tuple with
False and "This functionality is under
construction...come back later...". However,
the caller function, edit_startup_item, doesn't
do anything with this except store it in two
variables.

Also, I mentioned in the previous commit, 6a5bb8b,
that the signature of edit_startup_item was
changed. Unfortunately, due to how I staged and
then committed the changes, namely, breaking them
up over two different commits, the actual changed
signature is part of this commit.