Modified case 6 in the function edit_startup_item
Updated the call to the function save_startup_item
so it passed in the 3 required parameters instead
of the just the startup item dictionary.

Also added some print statements to show if the
json_reader function call in save_startup_item was
successful or not and to say that the save
functionality is still in progress and no data was
actually saved.