Stared working on function save_startup_item
Changed the parameter startup_item to
modified_startup_item. For now, added code to
grab the same startup item number from the JSON
data in general and print them out. Also added an
if statement to compare if there's a change and,
if not, return saying nothing was saved, but it
doesn't seem to work. Need to troubleshoot.