Fixed issue #9
Instead of adjusting the import statements in
demord.py, I adjusted them in startup_edit.py.
Now, all of startup_add.py is imported using a
regular import ... as ... statement. The module is
imported as a variable called dsa. Updated the
reference calls to the functions in startup_add.py
to use this variable. Tested that it worked.