Added menu option during new file creation to return to main menu
The menu for choosing to create a new startup file
with default values or user-provided ones didn't
allow to return to the main menu, so added that
option in. Changed the function new_file_chooser
in chooser.py to now return 2 boolean variables as
a tuple. Updated the main loop of demord.py to
account for this and only move forward if the user
chooses to create a new file.