Moved the default JSON data to a JSON file
Created a config folder and copied the default
JSON data from the DEFAULT_JSON variable to a
file called default_startup.json. The plan is to
use this file instead of a variable and read from
the file when needed. This will make it easier to
maintain what's used as the default JSON data.