Updated module startup_add
Had to add an import statement to import the
module demord specifically due to the order of the
import statements in the module imports.

In testing the save_startup_item function,
realized the if-block testing for a changed
startup item was testing for != instead of ==.
Changed that and tested it worked.