Created new prettify error function in module pretty
Created new function called prettify_custom_error
that takes in a string message and a string with a
function name. The following is then printed:

***ERROR***
The following error was encountered:

Function: <function_name>
Message: <string_message>

This function can be used at any point in the
program to provide error data that a user can
provide if they encounter the specific error.