Major changed to the function generate_user_edited_data
Updated the function to validate the data passed
in by checking that the modified JSON data has
all 7 keys that a startup item should contain, and
that if the item is to be added, that the original
JSON data has been passed in.

If either of these fails, an error message is
printed out and a blank dictionary returned.