Fixed some code in function generate_user_edited_data
Changed the error message printed when item_add is
False and the original JSON data dictionary is
blank. Had accidentally copied the same error
message as the other scenario so had to get the
initial message from commit 6ced24f from the
change log.

Also updated the if statement for the other
scenario, checking to see if the modified JSON
data is valid. Had to add in an or condition.