Major update to generate_user_edited_data
Refactored the function generate_user_edited_data
to properly validate all incoming parameters.
Updated the docstring to list the 3 different
scenarios to check for and coded that in.

Might need to utilise the jsonschema package in
PyPi to properly validate the variable
orig_json_data when it's supposed to contain full
JSON startup data.

For now, leaving the function as is.