Added a schema validator to module helper
Created a new function startup_data_validator. It
takes a dictionary called json_data and uses the
package jsonschema to validate against
startup_data.schema.json. If the JSON data doesn't
validate, an error is printed and the function
returns False. If the data validates, meaning the
passed in dictionary is properly formed startup
JSON data according to the schema, then True is
returned.