Refactored generate_user_edited_data for proper data validation
Using the helper module function
startup_data_validator, refactored the code to
validate parameter data and check for the 3
possible scenarios listed in the docstring.

Currently, multiple nested if-statements are used
to accomplish this with the necessary error
messages printing to the screen depending on when
data validation fails.

Need to clean that up and make the code more
efficient as much as possible.