Refactored the function data_validation_scenario
Rewrote the nested if-statements performing the
actual data validation into a match-case block.
Moved that match-case block to a new helper
function called match_scenario for easier code
reading, testing and maintaining.

The new function just takes the dictionary
data_validation and returns both an integer
representing the scenario number as well as an
error message if there is one. The function
data_validation_scenario still prints out the
actual error message.

For the match-case block, used the dictionary for
matching and matched based on the keys and their
possible boolean values. When 2 or 3 cases
resulted in the same error message, they were
merged into one case statement using the |
operator.