Created new helper function to aid with testing
In module startup_edit, created a helper function
called testing_shortcut_arguments_list. This lets
us the actual user menu and process when trying to
edit a startup argument.

Currently, the function
edit_startup_item_arguments_list checks to see if
in the module demord, the variable is_prod is
False. If so, the new helper function is called
and the rest of the
edit_startup_item_arguments_list function is
bypassed.

The new function takes the arguments list passed
in to the calling function and a boolean. A
boolean value of True means the helper function
will add an argument. A value of False means an
argument will be deleted.

This will specifically help with testing the save
functionality.