Made changes so the new helper function testing_shortcut_arguments_list works
Had to update the call to
testing_shortcut_arguments_list because in the
module demord, the is_prod variable cannot be
accessed since it's a local variable created in
the starting if block.

Made is_prod a global variable in demord and
added a helper function called is_production
which returns the global variable. This way, the
environment being used can be checked anytime.