Refactored data_generate to fix the issue mentioned in commit ef70fca
In commit ef70fca, a TODO comment was added to
the module data_generate for lines 183-207. This
was because in the function
data_validation_scenario, a KeyError was being
thrown due to the variable data_validation.

The issue was that the variable is a dictionary
but all the code utilising the variable was using
list referencing. This has now been fixed. The
keys in the dictionary were renamed and all
references to the dictionary for the functions
data_validation_scenario and match_scenario were
updated, including the match-case block in
match_scenario.