Used the changes made in commit c9e14d7 to fix an issue in validating the JSON data during the process to save a startup item
While testing the functions
data_validation_scenario and match_scenario in the
module data_generate, the print statements
modified and added in commit c9e14d7 revealed that
the validation was failing when it shouldn't.

Using the same print statements along with running
the Python debugger in VS Code, I realized that in
the module helper, in the function
startup_data_validator, the file path being used
wasn't complete because I assumed incorrectly what
the current working directory is. That is now
fixed.