Major update to function generate_user_edited_data in module data_generate
Added the functionality to handle scenarios 2 and
3 for generating JSON data that the user has made
changes to. As a reminder, the 3 scenarios listed
in the docstring of the function
data_validation_scenario are:

1. Need to update a single, existing startup item
2. Need to update the full JSON data
3. Need to add a single startup item

The function now handles when the data isn't valid
by returning an empty Python dictionary, and 2 of
the 3 scenarios. Only scenario 1 has to be coded
in.