Minor change to waiting for user input
In modules demord and startup_edit, there are 2
places where, after startup data is printed to
the screen, the program halts and waits for user
input to continue. There's also 1 place where the
prompt happens before the startup data is printed
out.

The prompts all used to start with
"Press any key", but it's only the enter key that
continues. Changed the prompts to start with
"Press the enter key when ready".