Refactored the if block in the function generate_user_edited_data
Realized that checking for the scenario where
validation fails wasn't necessary since an empty
Python dictionary is returned, and an empty Python
dictionary is created just before the if-block.

Adjusted scenario 2 to not return the variable
modified_json_data but to make a copy of it and
assign that to the new_json_data variable. This
way, there's only the one return statement at the
end.

Also, readjusted the if-block so it checks for
scenario 1 first, then scenario 2, then scenario 3
since that's more readable.