Refactored several testing related print statements
In functions match_scenario and save_startup_item,
refactored the code that printed out stuff for
testing purposes.

In match_scenario, the dictionary parameter passed
in, data_validation_ is printed out only if we are
in a testing environment. The function
is_production in the module demord is used for
this.

In save_startup_item, all the different print
statements were recoded to append to a string
which is then saved to a file called
save_fn_test.txt in the testing folder. This makes
it easier to compare and confirm that the code is
working.

The save code works except it doesn't currently
actually write anything to disk. Additionally,
while any changes made to the single startup item
being worked on persists, once the flow goes up
one menu level to select another startup item,
the changes made are lost. This should in theory
be ok as long as the save function is modified to
actually properly save data to disk.