Minor refactoring of the function edit_startup_item
In the menu presented for edit_startup_item, made
two changes so that instead of "View startup data"
and "Save startup data to disk", the choices now
read "View startup item" and "Save startup item
data to disk".

Also, updated case 6 for the match-case block,
to now print out the return message of the call to
the function save_startup_item when the status is
False.