Changed creation of startup file process to always use default data
Updated the generate_json_data function in module
data_generate. Now, when the user choose to create
a new startup file and chooses user-provided
values, the code checks the is_production function
in the main demord module.

If in testing environment, a message is printed
that the functionality isn't implemented and the
default startup data is being used. A call is then
made to the generate_default_startup_data function.

If not in testing environment, then the
generate_user_startup_data function is called like
normal.