Redid code changes made in a previous commit to skip the edit_startup_arguments_list function in the module startup_edit
The following code was added to the beginning of
the edit_startup_arguments_list function to skip
the whole function is in testing environment:

if not app_demord.is_production():
    print("Testing environment found: Skipping menu choice for shortcut")
    return testing_shortcut_arguments_list(arg_list.copy(), True)

The code was then commented out in a later commit.
The code is now uncommented so this function is
skipped for now.