Tested functionality to edit startup item arguments and to save the changes
Reverted the changes in the previous commit so
the function edit_startup_arguments_list in the
module startup_edit worked. Previous changes were
to skip this function when in testing environment.

Also tested the function by using item 2 from the
default startup data, which is to open Chrome
to Google. Added an argument, edited an existing
argument, and deleted an argument.

Finally, tested the save functionality mostly to
remind myself of the state of that functionality.
Forgot that the save function saves to a text file
in the testing folder the following, in order:

1. The original JSON data
2. The original startup item
3. The modified startup item
4. The new JSON data

The last part is done through the function
generate_user_edited_data in the module
data_generate.

Will need to go through the code to familiarize
myself with it. Will also need to test using
item 1 as outlined in issue #15