Started refactoring of code to use the function add_startup_item_arguments_list for every call to add a startup item argument
As written in the comments for issue12 on GitHub,
the module startup_add contains the function
add_startup_item_arguments_list. This function
should be used for every time a user wants to add
a new item argument. This includes both when
working with a startup item that has existing
arguments and one that doesn't.

Prior to this refactoring, the function
edit_startup_item_arguments_list in the module
startup_edit had code to directly take in a
new argument when the user wanted to add an
argument to a startup item with existing
arguments.

The code updated in this commit consists of
changes to the main if block in the
edit_startup_item_arguments_list function to
more easily show the code for each menu choice.
The block for adding an argument has a TODO
comment line and still needs to be actually
filled.