Completed step 2 of the necessary changes
In the comments for issue #15, the second step of
changes to make is:

2. Update the code listed in the previous comment
to remove the if arg_count >= 2 conditional

The reference is to the function
prettify_startup_item in the pretty module. Due to
the changes made in the previous commit to how
the ArgumentsList property is formatted, there is
no need for the nested if block checking for when
arg_count is equal to or greater than 2.

While testing the changes, encountered an error
where changes made to the last startup item would
cause the function generate_user_edited_data in
the data_generate module to fail. The issue was in
the if statement checking to see, when only one
startup item has been changed, which item number.