Resolved issue #17
Added a function called json_save in the jsonfn
module. Similar to the code for saving a single
startup item, this function calls the function
generate_edited_user_data in the data_generate
module.

Overall the json_save function allows
json_editor to have a save menu option to save the
full startup data. This means that a user can make
a change to a startup item and save just that item
in the menu for editing a single startup item. A
user can also make a change to one or more startup
items and then save them all at once.