Set up environment to test the startup.ps1 script with the changes made through demord.py
Copied startup.ps1 from the project root folder to
the branch root folder and renamed it to
test_demord-startup.ps1. Modified it to
automatically open the startup_data.json file in
the same branch root folder.

Tested using Windows Terminal and python.exe
making changes the startup file using demord.py
and then tested running the PowerShell script to
see that the changes work fine.

Everything was a success!

Noted one minor edit that could be made: some sort
of reminder to the user after they make a change
to item arguments to save, or, remove the ability
to quit the whole program from the edit/add item
arguments menu. I made the mistake initially of
changing an item argument for item 3 and then
quitting the program.