Added the ability to choose, when displaying a menu, whether to add a menu option to quit the program or not
Refactored the function user_menu_chooser in the
chooser module to take in an optional parameter
called allow_quit. It's a boolean that defaults to
true. When it's true, a final menu option to quit
the program will be shown for whatever menu is
being displayed. If it's false, this quit option
won't be shown.

Also reworded the menu option in the function
json_editor in the jsonfn module to save the full
startup data from 'save the current startup data'
to 'save full startup data to disk'. This
rewording mirrors the one to save a specific
startup item.