Renamed the function generate_json_data to generate_new_json_data in the data_generate module
Refactored that function to be solely for generating new JSON data. Removed the new_file parameter and also the **kwargs parameter, leaving only the is_default parameter.

Changed the code in the function as well as the call to that function from the jsonfn module.

This change allows the generate_user_edited_data function to be solely used for generating JSON data that the user edits.