Started modifying generate_user_edited_data as per commit 4537f8a
Changed the parameters for the function so that
the second parameter now can be used when adding,
deleting, or replacing a startup item.

Also added an if statement to valid the second
parameter. Will need to go through the rest of
the function code to add the deletion scenario and
also to adjust for the changed second parameter.