Changed the main if-block to a match-case block
Cleaned up the code a little by changing the if
statement in generate_user_edited_data that checks
the scenario number to a match-case statement.

May decide to create small helper functions for
processing each scenario because the code is still
a bit unwieldy and also an extra set of indents.