Initial commit for release v1.1-alpha
Modified the startup.ps1 file with the changes made in the branch features/startup_data_modifier_tool.

The changes consisted of refactoring the if block in the function Get-StartupItem because the schema for the startup JSON data changed. See commit eccc41b in the branch features/startup_data_modifier_tool for more details.

The same function was also updated to only grab the startup item properties that are currently used, such as ItemNumber, Path, ArgumentCount, and ArgumentList.

Will need to refactor the startup.bat file prior to release because it's currently hard-coded with the path to the startup.ps1 file.