Made further updates to the default startup data
Added the argument to the Chrome startup item to open in a new window.

Also updated the FilePath parameter for the Notepad startup item to use the full, absolute path over just "notepad". While this has worked in the past, it's not working now. So, just in case, it's better to use the absolute path.