Updated both startup PowerShell scripts to automatically work for both production and development environments
While testing things, realized that the changes made in commit 038dad1e caused the startup.ps1 file in the project root folder to not work properly since it was looking for a config folder.

Added an if-block to the startup script to check if the folder 'releases' is part of the script path by using the Test-Path cmdlet. Made the default folder path /config and then, if the cmdlet returns false, the script uses /data/json_data as the folder path.

Made these changes to the startup.ps1 file in the project root folder as well as the one in the /releases/v1/m1/Demord folder for release v1.1-alpha.

Finally, tested both scripts out using PowerShell and navigating to each folder respectively.