Refactored startup.bat to fix issue19
Followed the steps outlined in commit 06da807 and
refactored the startup.bat file to get the script
path and call the startup.ps1 file from the script
path.

Made the changes to both the startup BATCH file in
the project root folder and in
/releases/v1/m1/Demord for the release v1.1-alpha.
Tested that both work using cmd and navigating to
each folder respectively.