Moved the location of the instructions.txt file
After creating a ZIP file and then testing using
another profile the experience from downloading
release v1.1-alpha, realized that it makes more
sense to have the instructions.txt file outside
the Demord folder.

So, at the root level of the ZIP file will be the
instructions.txt file and the Demord folder.