Restructured the release contents
After doing a second test of the release ZIP file,
decided to make a package root folder that gets
zipped up.

For example, under /releases/v1/m1, there's now a
folder called Demord-1.1-alpha and inside that is
the instructions.txt file and the Demord folder.
Inside the Demord folder is startup.bat and
startup.ps1, as well as the config folder.
Finally, inside the config folder is the
startup_data.json file.