Minor updates to startup.ps1
While continuing testing of release v1.1-alpha,
realized that the startup PowerShell script still
checked for the old variable $IsProdEnv before
asking the user whether to run the script.

This was added when the variable was being used to
determine a testing or production environment.
However, with the more recent script changes, the
variable was removed. Refactored the script to
no longer check for that variable and just ask
every time.

Also added code to display if the user doesn't
enter "Y" or "N" to the question. It will now say,
"Please enter a valid choice". Made these changes
for both startup.ps1 files.