More minor updates to the startup PowerShell script
Added a printout of the message "Welcome to
Demord!" when the script runs. Also modified the
user input code to not use Read-Host so there's no
colon at the end of the prompt.

Rather, a Write-Host is used to show the prompt
text with a NoNewLine flag. This is followed by
a call to $Host.UI.ReadLine as found in this
Stack Overflow post:

https://stackoverflow.com/questions/50788152/read-host-always-ends-in-a-colon