Created a release notes folder for v1.1.-alpha
Inside releases/v1/m1, created a folder called release-notes. This
contains a Markdown file called gh_info.md which stores all the
information needed for the GitHub release of this version.

Initially, this version was released on GitHub, but then pulled
because the program/project name is being considered for change.
Once a new name is chosen and all branches updated accordingly,
the gh_info.md contents can be updated and used for a GitHub
release.