Moved startup.bat and startup.ps1 from the old Demord-1.1-alpha directory to the new CompStart-1.1-alpha directory
Along the way, also updated them as necessary to each have an appropriate first-line comment to the effect of "Main PowerShell/Batch script for CompStart".

In startup.ps1, made sure CompStart is used in the welcome message. Also, added a few spaces in between some lines so the code wasn't so bunched up.