Moved startup_data.json and updated the startup data
Moved the startup_data.json file from Demord/config to CompStart/config.

Also, realized that somwhere along the way, the startup data being used in this version in the releases directory was the old one with 7 startup items specific to my work.

Changed that to the default of 3 startup items.