Updated both startup scripts and the README in the feature_addons directory
Took the script changes made to startup.bat and startup.ps1 from commit 5a7a4abe and pasted them to the startup scripts in the devenv folder, so both versions will be the same.

Also, updated the feature_addons README to reflect the tool name change and the fact that the branch features/python_tkinter_test is now a past branch.