Deleted directory for feature branch features/startup_data_modifier_tool
Each feature branch used to only have its own directory in the feature_addons folder. However, when switching branches, git would report that it couldn't delete the folders from the other feature branches locally.

In a previous commit, after getting tired of being prompted to try again for the failed deletion attempts, I merged the feature branches startup_data_modifer_tool and this one together.

That meant both branches had the directories for both branches. I'm removing this because it's not how it should be!