Removed some commented out code in module startup_edit
In the function edit_startup_item_arguments_list, realized that formerly lines 192-197 contained commented out code. This mean was meant to skip the function for testing.

However, the code references an import of the main tool py file, which has been removed. Therefore, this commented out code was also removed.