Changed the default startup JSON data to be read in from the default_startup.json file
Updated the function generate_default_startup_data in the module data_generate to read in the JSON file with the default startup data and return it.

If the file doesn't exist where it should be, an error is printed out and a blank JSON object is returned.

If the file exists but there's a read error, that is printed out and a blank JSON object is returned.

Also updated the helper module and removed the DEFAULT_JSON global variable.