Fixed code after testing
Tested the changes made in the previous commit and received an error message.

Determined that due to the directory structural changes made to the CompStart project folder, the path needed to include "devenv".

Also, redid the code to generate the full path to be the same as in the main section of comp_start.py.

The function generate_default_startup_data now checks to see if we're in a production environment or not. If so, a Python list is created with "devenv", "data", and "json_data". If not, a Python list is created with "devenv", "feature_addons", "startup_data_modifier_tool".

This plus the filename is passed to the parse_full_path function in the helper module to be in line with the rest of the code that relies on this function to get a full path to a file.