Created new function in module helper
Created the function get_prod_path that can be used to get a starting location from which to look for files to open or read.

While working on this issue, after commit 518fe4c5, realized that the same if-block was repeated both in the main section of comp_start and in the generate_default_startup_data of the module data_generate.

This if-block was basically checking to see if the environment was production or not and then creating a list of strings with each one representing a folder, going from the CompStart project root folder to wherever the JSON files are stored.

Created this new function to do that, following the principle of DRY. Now, both spots in the code call this function and then append any other sub-directories as needed.