Fixed code after testing
After testing code changes in commit 7c921f3c, realized that for the scenario when not in production, the get_prod_path function wasn't accounting for some directory restructuring because of using an equals operator instead of calling the extend function for lists.

Fixed that and also fixed a spot in the function generate_new_json_data that was printing out the following message for when users wanted to generate both default and user created data:

"This functionality hasn't been fully implemented yet. Creating default startup file..."

Edited the if statement so that is only printed out when the user chooses to supply their own startup data, since generating the default startup data works fully but not user-generated data.

Finally, overwrote the existing startup_data.json file in the branch root folder, to test and confirm that the correct default data is being read in and saved.