Undoing deletion of text file relevant to issue #5
In commit 7cb9238d, the file commit_deets.txt, which was in the branch root folder for the branch features/startup_data_modifier_tool, was deleted.

According to the commit notes:

"Removed the file commit_deets.txt ... since this text file was used for an issue branch that has been resolved."

However, that is not true. This text file is relevant to issue #5, and so is still necessary for this current issue branch - dEhiN/issue5.

Therefore, I'm recreating the file using the same text as in the commit details.