Updated the line length
In VS Code, changed the Python Black formatter options to use 100 as the line length. This automatically was applied to lines of code in all Python files.

Also went through all docstrings and comments to apply the same line-length. The plan will be to update the line-length in the settings for all other locations where I work on this project, thereby keeping style consistency.