Refactored data_validation_scenario and started working on match_scenario
Based on the issue #5 comments in the Project board, the following 5 steps were determined to be next steps:

1. Refactor the function data_validation_scenario to utilize the item_type variable

2. Refactor the function data_validation_scenario to screen for all four scenarios, each of the four values of the item_type variable

3. Refactor the function match_scenario to account for the four scenarios

4. Decide if the function generate_user_edited_data will remove a startup item and update the startup JSON data, or if the calling function is responsible for that

5. Update all calls to generate_user_edited_data to pass in the correct parameters, mainly, that the item_type parameter is now a string

Note, the 5 steps are also listed in commit-deets.txt.

In this commit, step 1 is complete. Steps 2 and 5 are almost complete. The actual changes made to the modules data_generate and startup_add for steps 2 and 5, respectively, are most likely the only refactoring needed for the two steps, but just in case more sections of code are discovered, I will say steps 2 and 5 are in progress.

As a result of the current changes, match_scenario skips the actual matching and always returns scenario 0. This is so that I can test to make sure the correct data is being passed in.