Minor changes to the module data_generate
Removed the variable declaration for ENUM_ITV since it cannot be used directly, and a new helper function is going to be used to validate the item_type variable.

Also updated the key in the data_validation dictionary declared in the function data_validation_scenario. The first key was still called Item-Add, so changed it to Item-Type. It was already changed to have the value of the item_type variable passed in.