Created new helper function in module helper
Created the function check_item_type to take in the item_type variable used in the module data_generate. This is then compared, using a for loop, against all the member values from the ItemTypeVals class in the module enum. If a match is found, True is returned, otherwise False.

Also updated the function generate_user_edited_data in the module data_generate to call this helper function when validating the item_type variable.

While this helper function may only be called one in the code, at least so far, it's better programming to use a helper function since a single variable must be compared against all the members of an Enum class, and there's no single command that can do this.