Fixed problem in function json_data_validator that was causing validation to fail
Realized in the module helper, in the function json_data_validator, the full path to the JSON schema files hadn't been updated since the change to move all development files from the project root folder to /devenv.

Since the helper module already had a get_prod_path function from commit 7c921f3c, updated the function json_data_validator to call this function and then to append the "config" folder to what is returned.

This is the same behaviour followed in the function generate_default_startup_data in the module data_generate.

Will need to refactor this helper function and those that call it because currently, the path returned for the "production" environment is /devenv/data/json_data. However, once this tool is completed, that path will change.